requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
blosc2==2.0.0
certifi==2023.5.7
charset-normalizer==3.1.0
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.35
debugpy==1.6.7
decorator==5.1.1
earcut==1.1.5
executing==1.2.0
filelock==3.12.0
fonttools==4.39.4
frozenlist==1.3.3
h5py==3.8.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.23.1
ipython==8.13.2
ipywidgets==8.0.6
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
laspy==2.4.1
llvmlite==0.40.0
lz4==4.3.2
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
nest-asyncio==1.5.6
networkx==3.1
numexpr==2.8.4
numpy==1.24.3
numpy-stl==3.0.1
packaging==23.1
pandas==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
plyfile==0.9
pooch==1.7.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
PyFoam==2022.9
Pygments==2.15.1
pyparsing==3.0.9
pyproj==3.5.0
python-dateutil==2.8.2
python-utils==3.5.2
pytz==2023.3
pyvista==0.39.1
PyYAML==6.0
pyzmq==25.0.2
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
scooby==0.7.2
six==1.16.0
stack-data==0.6.2
sympy==1.12
tables==3.8.0
threadpoolctl==3.1.0
torch==2.0.1
tornado==6.3.2
traitlets==5.9.0
trame==2.5.0
trame-client==2.8.0
trame-components==2.1.1
trame-deckgl==2.0.2
trame-markdown==2.0.2
trame-matplotlib==2.0.2
trame-plotly==2.1.1
trame-rca==0.3.1
trame-router==2.0.2
trame-server==2.11.2
trame-simput==2.3.2
trame-vega==2.0.3
trame-vtk==2.4.4
trame-vuetify==2.2.4
trimesh==3.22.1
typing_extensions==4.6.1
tzdata==2023.3
urllib3==2.0.2
vtk==9.2.6
wcwidth==0.2.6
widgetsnbextension==4.0.7
wslink==1.10.2
yarl==1.9.2
zipp==3.15.0
aiohttp==3.8.4
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
blosc2==2.0.0
certifi==2023.5.7
charset-normalizer==3.1.0
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.35
debugpy==1.6.7
decorator==5.1.1
earcut==1.1.5
executing==1.2.0
filelock==3.12.0
fonttools==4.39.4
frozenlist==1.3.3
h5py==3.8.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.23.1
ipympl==0.9.3
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
laspy==2.4.1
llvmlite==0.40.0
lz4==4.3.2
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
nest-asyncio==1.5.6
networkx==3.1
numexpr==2.8.4
numpy==1.24.3
numpy-stl==3.0.1
packaging==23.1
pandas==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
plyfile==0.9
pooch==1.7.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
PyFoam==2022.9
Pygments==2.15.1
pyparsing==3.0.9
pyproj==3.5.0
python-dateutil==2.8.2
python-utils==3.5.2
pytz==2023.3
pyvista==0.39.1
PyYAML==6.0
pyzmq==25.0.2
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
scooby==0.7.2
six==1.16.0
stack-data==0.6.2
sympy==1.12
tables==3.8.0
threadpoolctl==3.1.0
torch==2.0.1
tornado==6.3.2
traitlets==5.9.0
trame==2.5.0
trame-client==2.8.0
trame-components==2.1.1
trame-deckgl==2.0.2
trame-markdown==2.0.2
trame-matplotlib==2.0.2
trame-plotly==2.1.1
trame-rca==0.3.1
trame-router==2.0.2
trame-server==2.11.2
trame-simput==2.3.2
trame-vega==2.0.3
trame-vtk==2.4.4
trame-vuetify==2.2.4
trimesh==3.22.1
typing_extensions==4.6.1
tzdata==2023.3
urllib3==2.0.2
vtk==9.2.6
wcwidth==0.2.6
widgetsnbextension==4.0.7
wslink==1.10.2
yarl==1.9.2
zipp==3.15.0
